revert to bind handler selectors
This removes the `pattern` field from Rule and associated logic from the
Job type. This does retain the input directory walking and collecting of
paths which is then projected into the Bind extensions for the benefit
of certain Bind handlers such as `Select` which takes on the role of the
Pattern field of Rule.