`Site`'s configuration is no longer immutable
This was unnecessary and complicated the APIs. Instead it will be
possible to continue to mutate/configure the `Site` configuration, and
it will only be copied and frozen into an `Arc` when the site is built.

This also streamlines the `Command` API because now all commands will
simply receive a `Site` and can continue to configure it as they want,
e.g. the `live` command can configure the `Site` to turn on the
`is_preview` flag.

I'm still not entirely sure that there should even be a `Configuration`
type with blessed configuration values, or that it should be stored
separately instead of within the `Bind` extensions.