implement new Command API
Things are much simpler now. The reason for doing this was to enable the
configuration of commands and to avoid the separate stages of
configuration.

This got rid of the Plugin wrapper type and Plugin constructor
gymnastics. Instead now there's just the Command trait with run and
description methods (still not sure about the latter). The run command
takes a mut ref to the Site, which can be further configured by using
the configure_mut() method on Site. A common pattern that I've been
using is to define a configure() method on the command type and then
call that at the beginning of the Command::run method impl.

This has considerably simplified things.