migrate away from eventual to simple thread pool
Now jobs are processed in batches, so that the current set of ready jobs
are process completely, then the results are processed and the new ready
jobs are enqueued and the process repeated.

I originally preferred to be as parallel as possible, in the sense that
jobs would process as soon as they were ready.

However, I think this new way of batching jobs to run in parallel is
closer to what cargo does and it seems to be pretty fast and much easier
to follow, so I think for now I'll keep it this way and perhaps revisit
it if I come up with a better idea.